requirements.txt
customtkinter==5.2.2
matplotlib==3.10.1
numpy==2.2.4
pandas==2.2.3
Pillow==11.1.0
PyMeasure==0.15.0
